verify result in ReturnExecutorTest and move *Stats to "Aql/Stats.h"
